Legacy: Fixes virtual keyboard on BlackBerry
If `virtualkeyboard_change_options()` is not called, whenever the user requests the virtual keyboard the last one used by the operating system will be displayed. For instance, if the user unlocks the phone (numeric layout), goes directly to a Lime / OpenFL application and requests the keyboard, the one displayed will be the numeric one.