Several changes to the XCode project template to fix warnings
Adds flags for testability, only to build active architecture, and bakes the app package name into the project file. All this to fix the warnings generated by XCode.
This also drops the flags for architecture completely as there is only one valid variation for tvOS and not setting it gives the default.