Restore the ability to look up a directory.
This time I'm using the string "directory" rather than "cwd", "pwd", and "workingDirectory". The first is a reference to Haxe's `Sys.getCwd()` function, which is how HXProject originally determines the value of `projectDirectory`. And the others are references to the same concept.

However, part of the point of this pull request is that you can look up a parent project's directory. When you do, that directory will no longer be the working directory, so all three of these identifiers are misnamed.

I considered "projectDirectory", but I think a single word works best.