Add batch event processing in lime/sdl
Provide the BatchUpdate(n) method in the SDLApplication class
that supports pulling multiple events at a time from the sdl
event queue. BatchUpdate does similar work to Update, but it
will never wait indefinitely for an event to become available,
and it accomplishes its work more efficiently. Events are
fetched to a buffer at the class level, and this buffer needs
to be freed in ~SDLApplication().

Because BatchUpdate will never wait indefinitely, it can be
called in Init to process all pending events at startup (they
can accumulate) and to initialize the framerate timer for targets
that make use of it. This allows us to clean up Update by
removing the firstTime boolean in its event loop.

Going forward, BatchUpdate is a tool that can be used to
refactor the event loop logic, improving it for all targets. It
can also be used to implement the event loop inversion of control
demonstrated at Haxe Summit 2019 for embedding OpenFL apps into
the windows of host applications.