Fix CI edge case
Depending on how the CI server is set up, the task may fail to
create the `lime` command and stop there. For instance, #1416
returned failure due to dependabot's settings, not due to the
contents of the pull request.

This commit provides a simple workaround for those cases, running
only if `lime setup -alias` fails. It shouldn't have any effect in
cases that already work.