Allow hyphen for launchStoryboard template property typing
When parsing the attributes attached to a launchStoryboard element, in addition to the colon character (`:`), allow the hyphen character (`-`) to be used as the separator between a template property's name and its type.

Although colon is familiar to Haxe developers as the variable type specifier (e.g. `var i:Int = 0;`), the colon character's expected usage within xml is to separate the local name from the namespace. So usage of colon here could be unexpected, and also causes trouble in formalizing the specification as an xsd.

Retain the ability to parse `:` for backward compatibility.