Fix `NativeAudioSource.stop()`
The `setCurrentTime()` function calls `AL.sourcePlay()` if `playing` is
still true, so that should only be called after `playing = false;`.

One could also place the call before `AL.sourceStop()`, but that will
lead to a few extra function calls. Plus, if `stop()` is supposed to
reset the playback time, it should probably do that whether or not the
sound is playing (i.e., outside the `if` block).