Make `sendComplete()` more thread safe.
Modifying `complete` on the worker thread can lead to inconsistencies. A
function on the main thread might check `if (!worker.complete)` and then
run some code. But the background thread could in theory set `complete`
right after the main thread checks it. Then you're executing the "not
complete" branch of the if statement even though `complete` is true.

This one is minor, but you have to be careful about threads.