Only run `BackgroundWorker`s once.
This was always the intended use case, and the alternative produces all
kinds of subtle issues. Even with a great deal of effort, it isn't
possible to account for all the little ways one thread could cross over
to affect another.

It's still possible for an old thread to call a new `BackgroundWorker`'s
`sendComplete()` function (among others), but now this will be obvious.