Consolidate paired messages into one.
It's theoretically possible for the OS to delay a thread at any time,
even between adjacent calls to `__messageQueue.add()`. This means the
second half could be missing when the main thread goes looking for it,
and would be lost forever given how `__update()` works.

Even if that's super unlikely, it's still safer to send single messages.