Implement `restoreInstanceMethods()`.
Another annoying feature of JavaScript, another workaround.

An alternate implementation would be to automatically apply this to all
function arguments, iterating through anonymous structures as needed.
It'd be tricky, but not impossible. The problem is that there would be
no way to pick and choose which objects were affected, making it extra
hard to track down errors.