Use the newer Xcode build system introduced in Xcode 10
Up until now the legacy Xcode build system has been being used.
As of Xcode 14, the legacy build system has been removed.

In order to keep products being built to a known location local to
the project folder, the derived data path is passed to all calls to
xcodebuild. This also requires the desired scheme to be passed.
Previously, scheme was only passed for archive operations, but now
it's a common argument for all calls to xcodebuild.

Without specifying the desired derivedDataPath, it would end up being
"~/Library/Developer/Xcode/DerivedData/{generated-project-id}", and
xcodebuild would need to be queried for the exact path.