Fix Main.cpp template for -static -debug builds
When building a Lime application in debug mode on Windows, the console
subsystem is used and Windows looks for a 'main' function. However, the
Main.cpp file used when linking statically always defines a 'WinMain'
function regardless of whether the application is being built in debug
mode.

This commit adds an additional check in the Main.cpp that defines a
'main' function instead of 'WinMain' when building in debug mode.