AudioSource position changes
Position shouldn't be applied by default, since it's extra code
running that may not be needed, and in some cases can significantly
alter the sound quality. Specifically, spatializing sounds on HTML5
can make sounds more muffled.

This change caches the position in AudioSource, and all getPosition
functions in the AudioSource backends are replaced with accessing
the cached position value.