Use `implementation` instead of `api`.
This is recommended as a best practice, though AFAIK this only matters
for projects that are going to be imported by other projects.

For instance, you used to be able to include `:deps:extension-api` by
including any extension that depended on it. Now, every project that
wants to use `extension-api` has to include it directly. (Which is fine
because in practice, they all already do so.)