NativeWindow: Implement Window Styles
SDL2 either supports(or is in the process of) window styling that allows developers to further customize window behavior.

SDL does implement window flags with no taskbar on both Windows and X11 already, which will allow developers to apply pop-ups or custom context menu's to their applications by using such a feature. I think it should be considered for a future Lime release.