Fix accidental HTML tag.
The tag doesn't get rendered, and it looks like the line says "Fixed
merging of multiple tags in project files." Which makes perfect sense at
a glance, so it isn't surprising this wasn't noticed.

Thanks to markdownlint for catching this. Now that all its minor
warnings are gone, important ones will stand out.