Improve "Could not find NekoAPI interface" message.
This message comes up relatively often when a new user tries to set up Lime, but fails to give any instructions a new user could use.

This commit adds a new message explaining the most common issue and how to solve it. It also provides file paths, which may help with more in-depth debugging.