Rewrite `PlatformTarget.execute()`.
This function attempts to skip steps if the corresponding function was not overridden. However, in `-eval` mode, it may incorrectly detect the `@ignore` metadata on subclass functions, causing it to always skip.

Running an empty function is not a problem, but we wouldn't want to log that step if nothing was being done. Therefore I switched the check to only apply to the log call.

For one reason or another, the "watch" command can't be skipped, so I left it alone.