CI: use `-eval` to run air-samples.
This allows it to run immediately, rather than having to wait for Windows. Plus, it makes the workflow summary clearly legible again.

The downside is that `-eval` wasn't available in Haxe 3, so we can no longer test that. However, I just added more Neko+Haxe 3 tests, so I think it balances out.