NativeWindow: call setTextInputEnabled(false) right away when window is created (closes openfl/openfl#2697)
We seem to have been assuming that it was false by default, but in SDL 2, it is true by default (planned to be false in SDL 3, apparently). It can cause weird behavior like IME popups appearing when holding down keys (something commonly done by users playing games, so we don't want that!). We want setTextInputEnabled(true) to happen only when an OpenFL TextField (or anything else that accepts text input) receives focus, and then cleared again when focus is lost.

This does not disable regular keyboard input. It is specifically related to what SDL considers text input, which is different, even if it uses the keyboard.