FileDialog: Fixes issue where file dialog stops responding in HashLink on Windows (closes #1849)
Does not seem to recover when it happens, requiring the Lime app to be force closed. Some kind of threading issue that is affecting HashLink on Windows only (macOS and Linux are fine, and other sys targets are fine).

Fixed by passing SINGLE_THREADED to ThreadPool constructor using #if (windows && hl)