Revert change to Sys.setCwd behavior
Before `Sys.setCwd()` would only be run in the `rebuild tools` case,
where it would always been cleaned up. The previous patch changed it so
that Sys.setCwd runs for all cases, even if it won't be cleaned up. This
patch corrects that error, and once again makes sure that setCwd is only
run for `rebuild tools`.